+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### VNEXT
 
-* Added support for custom Directives [Issue #212](https://github.com/apollographql/graphql-tools/issues/212) [PR #518](https://github.com/apollographql/graphql-tools/pull/518)
+* Added support for custom Directives on FIELD [Issue #212](https://github.com/apollographql/graphql-tools/issues/212) [PR #518](https://github.com/apollographql/graphql-tools/pull/518)
 
 ### v2.11.0